Update LivelinkConnector.java
There was a bug in this file since version 1.8, when we try to add a Livelink folder in Path tab, it was break and today I found that this was because, you missed seqPrefix+ in line no 3265. Please consider changing this in the next release.
Now since, I became more familiar with ManifoldCF architecture, I may be try to find more bug in livelink connector here on.

Thanks for such a great tool to index.