Add non-emptiness clause to realm check of Solr output connector
The underlying HTTP client of the Solr output connector always configures the AuthScope of the connection, even if the provided realm is the empty string (""). This is due to the fact that the attribute realm is never null, but "" in case the user did not supply a parameter for this. This breaks connectivity in case basic auth is in-place, but no realm is required. Thus, in both cases (null value and "" for realm) the HTTP client should use AuthScope.ANY. This commit adds the necessary non-emptiness clause to the realm check in order to fix this behavior.